Move Search Parser to a dedicated dir and ignore it for typecheck GH workflow
