+++ src/libs/SearchParser/searchParser.js
